(nw) softlist_dev.cpp: validate ROM labels
The free-for-all on labels in software lists is not working.  There's no
consistency, labels are getting excessively long, people are starting to
use non-ASCII characters in labels making it harder for others to type
them when manipulating files on the command line, and there's too much
markup being put in labels.

The length limit is 127 characters, same as for labels in MAME itself.
This should be long enough to be descriptive.  Remember that the Win32
path limit is 260 characters, and many applications and frameworks have
issues with longer paths, including Windows Explorer and the .NET
framework.  Labels are used as filenames, so concessions need to be
made for this.

I have not abbreviated excessively long labels myself - they're
currently causing 135 validity errors.  Someone else can fix them.

Printable ASCII characters are allowed, with a few exceptions.  The
exceptions are limited to characters most likely to cause issues for
interactive shells and scripts:
* ! - csh event substitution (very difficult to escape properly)
* $ - sh varibale expansion
* % - csh job control, cmd variable expansion
* / - UNIX directory separator
* : - sh path separator, Windows drive qualifier
* \ - sh escape, Windows directory separator

Most of the labels that had to be edited were using ! for markup, or
using ! and % for titles in labels.  Strangely, titles in labels are
often forced to lower case, despite this never being enforced for
software lists.  There are also various other edits to titles used for
labels, such as moving articles to the end (with or without a comma),
or replacing spaces with underscores.  As I already said, there's no
consistency at all.

There is far too much markup in labels.  They're even being used for
notes in some cases (e.g. at least one case where a dumper's name is in
the label).  The XML schema supports metadata - use it.  For example,
you can use part_id for an unrestricted display name for a software
part.  You can also use XML comments for notes.

And while on the topic of metadata, vgmplay.xml is putting the same
thing in the part_id as well as the label.  The part_id should have
the actual title, not the title mangled to make it more suitable for
use as a filename.  Addressing this would be a lot of work, given how
large the file is.

For now, empty data areas in software lists cause a verbose message
rather than a validation warning.  There are thousands of software
lists using empty data areas to indicate the size/width of cartridge
RAM/EEPROM/etc.