netlist: reduce usage of NETLIB_CONSTRUCTOR_DERIVED. (nw)
Reduce the usage of NETLIB_CONSTRUCTOR_DERIVED to implement
dip package objects. This aligns the programmatic approach
with the approach used in the macro folder, i.e. using netlists.

Untangle model definition and interpretation as well.

This changes are a first step in aligning models and families and thus
simplify the source further.