netlist: further optimize ppmf.h [Couriersud]
Move towards a pure c++ constexpr solution to derive the optimal
pointer to member function solution.

All the macro madness is centralized to a compile_info struct with
static members.

For the time being the evaluation result is compared against the
previous approach. Going forward this will be dropped as well as the
support for MINGW with GCC <= 4.6