drcbex64: use asmjit
(nw)This is essentially a 1-for-1 substitution of x86emit code with its asmjit equivalent. There are opportunities for refactoring some of the original code using C++ as well as asmjit features, but this first pass simplified identifying and correcting any problems.

Selective tests with mips3, hyperstone and sh4 appear to reproduce existing behaviour but more testing is needed.