Pass FORCE_C_DRC_BACKEND=1 on ppc architectures
PowerPC architecture is special in a way that it has some inline assembly
code but no DRC support. As a result, NOASM=1 was never configured.
In contrast to the old DRC, asmjit only compiles on the architectures it
supports. FORCE_C_DRC_BACKEND=1 needs to be passed to the makefile or
the compilation will fail.