VG5000µ update:
- Add support for the "Delta" key

The "Delta" key on the VG5000µ keyboard is wired to directly assert the NMI line of the Z80. The handler for the NMI in the BASIC ROM checks if the CTRL key is also depressed and then causes a soft reset.

This routine can be hooked to produce other effects than soft reset.