-emu/rendlay.cpp: Fixed bug calculating group bounds with repeats.
Bounds were being miscalculated when the first element inside the group
was inside a repeat or a collection.  Most visible in md6802.

-Updated another batch of layouts.

A lot of these were using a black backdrop element to force bounds when
either explicit view bounds or an empty element is a better way to
achieve the same thing.

Bally by17/by35 pinball layouts have had a bit of an overhaul with the
three sections of the display now abled to be toggled independently.

Ensoniq VFD layouts refactored to reduce a very large number of lines
down to a few nested repeats.