Cleaned up inputs a little.
Removed the IPT_POKER_BET, replacing it with IPT_GAMBLE_BET.

Re-organised the order of input types in the source and the
documentation to better reflect how they're logically grouped or
organised on a mahjong or gambling control panel.

Made the blackjack, poker and slots documentation headings subheadings
under the gambling heading.

Cleaned up inputs for a number of gambling games.