machines promoted to WORKING (Plug and Play) (#7282)
machines promoted to WORKING
-----
Golden Nugget Casino [David Haywood, Sean Riddle]
Sudoku Plug & Play TV Game '6 Intelligent Games' [David Haywood, Sean Riddle]
Frogger (MSI Plug & Play, white joystick) [David Haywood, Sean Riddle]
35 in 1 Super Twins [David Haywood, Sean Riddle]
Vs. Maxx 17-in-1 [David Haywood, Sean Riddle]

made use of the code that was already partially present in the NES PPU code to handle emphasis modes, used to highlight tally bars etc. in 'nes rampart', apply a red tint to 'fds bublbobl' and dim the screen in a number of other games.

fixed a number of issues with VT palette modes that was causing black screens in many games, allowing them to be promoted

added a timing kludge for interrupts in the NES Frogger titles, until timings are better overall,makes playfield stable here and in a number of other places so that other things can be improved

note "Vs. Maxx 17-in-1" was added as NOT WORKING in this same cycle so if this gets merged before freeze they'll need to be moved to the 'new working' section.

This is a checkpoint, further code improvements and refactors will come later.