msm5232: fix: behaviour when the 'arm' flag set while envelope generator is on a decay state
msm5232, a tone generator chip which is used by some early TAITO and Alpha Denshi arcade PCBs, has 'arm' flag that alters behavoiur of envelope generation.

Basically, the 'arm' flag turns Attack-Decay-Release envelope generation to simpler ON/OFF states.
If the 'arm' flag set, the transition from Attack to Decay doesn't occur unless explicitly indicated.

On current implement, when this 'arm' flag is set on a channel which is on a Decay state, the Decay state continues.
It seems that, the state should be turned into a Attack state in this situation.