HD61830: Fix character blink mode
When Cursor=Off and Blink=On, the controller is in "Character Blink"
mode, according to the data sheet.

This means that the character cell at the the position of the cursor is
turned fully ON at a 50% blink duty cycle. Verified on a device with a
real HD61830 (Atari Portfolio).