machine/latch8.cpp: Clean up a little.
Making the bit read functions honour the XOR mask is easy in theory, but
dkong.cpp is doing scary things with the XOR mask and bit read handlers
on one of its latches, so it could end up breaking something.  Honouring
the read callbacks would add two tests to each of the bit read functions,
which could hurt performance.  The whole design of this device seems
somewhat incoherent.