cpu/drcbex64.cpp: Proof-of-concept for optimisation of calling out.
Resolve address space virtual member function addresses on constrcution
and call them directly.  Provides a small but measurable improvement to
performance in drivers that use the recompiler and access the memory
system a lot.

Also made MSVC delegates capable of walking past all the thunks MSVC
puts in the way of actually calling a member function.  I'm not
accounting for the "this" pointer being passed in RDX when the return
value is an oversize scalar.  This is harmless because it won't see
anything that looks like a virtual call thunk using RCX when RCX points
to uninitialised space for the return value.  It just means virtual
member function calls won't be bypassed if the return value is an
oversize scalar, but that doesn't happen frequently anyway.