More UI enhancements/cleanup:
Made the headings in the info viewer clickable, so you can switch
between DATs with a mouse or trackball (or maybe a lightgun if that's
your thing).

Made the UI red/yellow/green traffic light status colours less
dirty-looking.  The "yellow" is more of an amber-brown than a cat puke
brown now, and red is brighter.  The contrast with white text is
definitely fine for red and green, but it's reduced a bit for yellow.
However there's a limit to how dark you can make a yellow or orange
colour before it looks muddy.

Reduced the number of places the UI uses red for things that aren't
errors.  The error colour should be used sparingly, for actual errors.

Improved the colour swatch display in the RGBA colour editor.  It now
has black/white underlay so alpha effects are move obvious.  Also fixed
a bug preventing the channels being run down to zero.

Fixed double-clicking configure machine causing MAME to exit from the
system selection menu.  Also slightly adjusted the colours of the
toolbar buttons again.