frontend: Beter provision for plugins with multiple menu levels.
Make it possible for a plugin to specify which item should be selected
on repopulate.  Also allowed plugins to specify the UI heading flag for
menu items.

Updated input macro plugin to select sane items when changing menus.
Other plugins not updated yet, but may be at some point in the future.