Updates to mhavocpex* drivers
Optimized drivers.
Added Cocktail support which is now in the new software and was on original hardware but was never implemented in MAME because the original software did not use it.
Verified that both LPC and ADPCM version of mhavocPEX are working.
Updated inputs for support of cocktail mode spinners.