machine/dp83932c.cpp: Support BE mode and transmission packet appending (#9910)
Added support for big endian bus accesses by adjusting the width of the bus access appropriately.  (In 32-bit mode, descriptors and the RBA must all be aligned to 32-bit boundaries, even when the values are only 16 bit, thus the previous method had issues on BE buses.  See pages 7 and 15 in the datasheet.)

Stop device from starting a second transaction task if one is already in progress.  The datasheet recommends that software which wants to append packets during a transmission send another transmission command to avoid a race condition, and it should do nothing if this command is received while the transfer is still ongoing.  NEWS-OS 4.2.1aRD is an example of software that does this.  It would hang on boot without this change because the SONIC would start doing weird things when the second transmit command came in.